chart editor example updates
