+++ vegafusion-server/examples/chart_editor/index.html
@@ -1,43 +1,54 @@
 <!DOCTYPE html>
 <html>
-  <head>
-    <meta charset="utf-8">
-    <title>Hello wasm-pack!</title>
-  </head>
-  <body>
-    <div id="vega-chart", class="vega-chart"></div>
-    <div>
-      <ul class="nav nav-tabs">
-        <li class="nav-item">
-          <a href="#editor" class="nav-link active" data-bs-toggle="tab">Editor</a>
-        </li>
-        <li class="nav-item">
-          <a href="#server-spec" class="nav-link" data-bs-toggle="tab">Server</a>
-        </li>
-        <li class="nav-item">
-          <a href="#client-spec" class="nav-link" data-bs-toggle="tab">Client</a>
-        </li>
-        <li class="nav-item">
-          <a href="#comm-plan" class="nav-link" data-bs-toggle="tab">Comm Plan</a>
-        </li>
-      </ul>
-      <div class="tab-content">
-        <div class="tab-pane show active" id="editor">
-          <div id="editor-monaco" class="monaco-editor"></div>
-        </div>
-        <div class="tab-pane" id="server-spec">
-          <div id="server-spec-monaco" class="monaco-editor"></div>
-        </div>
-        <div class="tab-pane" id="client-spec">
-          <div id="client-spec-monaco" class="monaco-editor"></div>
-        </div>
-        <div class="tab-pane" id="comm-plan">
-          <div id="comm-plan-monaco" class="monaco-editor"></div>
+<head>
+  <meta charset="utf-8">
+  <title>Hello wasm-pack!</title>
+</head>
+<body>
+
+<div class="container">
+  <div class="col">
+    <h1>VegaFusion</h1>
+    <div class="card">
+      <div class="card-body">
+        <ul class="nav nav-tabs">
+          <li class="nav-item">
+            <a href="#editor" class="nav-link active" data-bs-toggle="tab">Editor</a>
+          </li>
+          <li class="nav-item">
+            <a href="#server-spec" class="nav-link" data-bs-toggle="tab">Server</a>
+          </li>
+          <li class="nav-item">
+            <a href="#client-spec" class="nav-link" data-bs-toggle="tab">Client</a>
+          </li>
+          <li class="nav-item">
+            <a href="#comm-plan" class="nav-link" data-bs-toggle="tab">Comm Plan</a>
+          </li>
+        </ul>
+        <div class="tab-content">
+          <div class="tab-pane show active" id="editor">
+            <div id="editor-monaco" class="monaco-editor"></div>
+          </div>
+          <div class="tab-pane" id="server-spec">
+            <div id="server-spec-monaco" class="monaco-editor"></div>
+          </div>
+          <div class="tab-pane" id="client-spec">
+            <div id="client-spec-monaco" class="monaco-editor"></div>
+          </div>
+          <div class="tab-pane" id="comm-plan">
+            <div id="comm-plan-monaco" class="monaco-editor"></div>
+          </div>
         </div>
       </div>
     </div>
-
-    <noscript>This page contains webassembly and javascript content, please enable javascript in your browser.</noscript>
-    <script src="./bootstrap.js"></script>
-  </body>
+    <div class="card">
+      <div class="card-body">
+        <div id="vega-chart" class="vega-chart"></div>
+      </div>
+    </div>
+  </div>
+</div>
+<noscript>This page contains webassembly and javascript content, please enable javascript in your browser.</noscript>
+<script src="./bootstrap.js"></script>
+</body>
 </html>
