+++ vegafusion-server/examples/chart_editor/style.css
@@ -1,14 +1,3 @@
-html,
-body {
-    margin: 0;
-    padding: 0;
-    height: 100%;
-}
-
 .tab-pane .monaco-editor {
-    height: 500px;
-}
-
-.vega-chart {
     height: 500px;
 }
