+++ vegafusion-wasm/package.json
@@ -38,6 +38,7 @@
     "lodash": "^4.17.21",
     "vega": "5.21.0",
     "vega-tooltip": "^0.27.0",
-    "vega-util": "^1.17.0"
+    "vega-util": "^1.17.0",
+    "grpc-web": "^1.3.1"
   }
 }
